fix func tests
