+++ chart/templates/pohttp-svc.yml
@@ -11,8 +11,8 @@ metadata:
 spec:
   type: {{ .Values.service.type }}
   ports:
-    - port: 8080
-      targetPort: 8080
+    - port: 80
+      targetPort: pohttp
       protocol: TCP
       name: http
   selector:
